food_order_assistant/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
tqdm
openai
sentence-transformers
elasticsearch
psycopg2-binary